修复了两个警告问题 添加了对printf风格的参数校验支持
修复了两个警告问题 添加了对printf风格的参数校验支持

+++ src/u8g2_menu.c
@@ -538,11 +538,11 @@ void u8g2_MenuItemMove(u8g2_menu_t *u8g2_menu, u8g2_uint_t i)
 
 void u8g2_MenuEvent_itemBindChar_itemMove(u8g2_menu_t *u8g2_menu, void* value)
 {
-    u8g2_MenuItemMove(u8g2_menu,(uint8_t)value);
+    u8g2_MenuItemMove(u8g2_menu,(uint8_t)(uint32_t)value);
 }
 void u8g2_MenuEvent_itemBindChar_itemSelect(u8g2_menu_t *u8g2_menu, void* value)
 {
-    if((uint8_t)value)
+    if((uint8_t)(uint32_t)value)
     {
         u8g2_MenuItemSelect(u8g2_menu);
     }
